requirements.txt
deepgram-captions==1.0.5
deepgram-sdk==2.12.0
pytube==15.0.0
streamlit==1.28.1
streamlit-player==0.1.5
python-dotenv==2.8.1